Move extended PATH into build.sh.
When setting this from inside the makefile the PATH will change during
processing of the makefile. This makes it hard to detect installed
compiler versions, etc. Now we will still have the non-extended path
during execution of make print_extended_path but during this nothing
that depends on the PATH is executed so this is fine.